feat(batcher): make configuration explicit and document commit semantics
A batcher has always required either a size limit or a timeout to be
configured. Given this requirement, using explicit parameters is clearer and
more appropriate than the functional options pattern when instantiating a
batcher.

Improve the overall documentation, particularly by clearly defining the
requirements for commit functions and the guarantees provided around their
parameters and the commit process.

Closes #34